+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added a possibility of adding own isotopes.
 - Dipole-dipole coupling constants and tensors can now be requested for a spin
 system.
+- Quadrupole coupling tensors can now be requested for a spin system.
 
 ## 0.1.1
 - Added quadrupolar moments to several isotopes.